Revert "Simplify adding icons to the desktop"
This reverts commit e339463f95c70578aae33f339ac54433f51cc5ae.

This patch had the added feature of crashing xfdesktop when the
icon size shrinks. Reverting to clean up the patch.

+++ src/xfdesktop-file-icon-manager.h
@@ -59,8 +59,8 @@ gboolean xfdesktop_file_icon_manager_get_cached_icon_position(
                                                     XfdesktopFileIconManager *fmanager,
                                                     const gchar *name,
                                                     const gchar *identifier,
-                                                    gint *row,
-                                                    gint *col);
+                                                    gint16 *row,
+                                                    gint16 *col);
 
 G_END_DECLS
